Detectando codificacion de archivos
 Please enter the commit message for your changes. Lines starting

+++ app/src/ConfigFont_class.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 
 
-    size_t ConfigFont_class::found = (-1);
+    int ConfigFont_class::found = (-1);
     string ConfigFont_class::p1 = "";
     string ConfigFont_class::p2 = "";
     int ConfigFont_class::p3_Height = 0;
@@ -13,5 +13,7 @@ using namespace std;
     int ConfigFont_class::p5_maxLinLen = 0;
     int ConfigFont_class::p6_defSmuMod = 0;
     int ConfigFont_class::p7_numComm = 0;
+
+
